# Move index values cache to bucket level to prevent race condition