--- a/pkg/block/indexheader/binary_reader.go
+++ b/pkg/block/indexheader/binary_reader.go
@@ -418,8 +418,6 @@ type postingOffset struct {
 	tableOff int
 }
 
-const valueSymbolsCacheSize = 1024
-
 type BinaryReader struct {
 	b   index.ByteSlice
 	toc *BinaryTOC
@@ -438,12 +436,6 @@ type BinaryReader struct {
 	// Cache of the label name symbol lookups,
 	// as there are not many and they are half of all lookups.
 	nameSymbols map[uint32]string
-	// Direct cache of values. This is much faster than an LRU cache and still provides
-	// a reasonable cache hit ratio.
-	valueSymbols [valueSymbolsCacheSize]struct {
-		index  uint32
-		symbol string
-	}
 
 	dec *index.Decoder
 
@@ -646,12 +638,12 @@ func newBinaryTOCFromByteSlice(bs index.ByteSlice) (*BinaryTOC, error) {
 	}, nil
 }
 
-func (r BinaryReader) IndexVersion() (int, error) {
+func (r *BinaryReader) IndexVersion() (int, error) {
 	return r.indexVersion, nil
 }
 
 // TODO(bwplotka): Get advantage of multi value offset fetch.
-func (r BinaryReader) PostingsOffset(name string, value string) (index.Range, error) {
+func (r *BinaryReader) PostingsOffset(name string, value string) (index.Range, error) {
 	rngs, err := r.postingsOffset(name, value)
 	if err != nil {
 		return index.Range{}, err
@@ -674,7 +666,7 @@ func skipNAndName(d *encoding.Decbuf, buf *int) {
 	}
 	d.Skip(*buf)
 }
-func (r BinaryReader) postingsOffset(name string, values ...string) ([]index.Range, error) {
+func (r *BinaryReader) postingsOffset(name string, values ...string) ([]index.Range, error) {
 	rngs := make([]index.Range, 0, len(values))
 	if r.indexVersion == index.FormatV1 {
 		e, ok := r.postingsV1[name]
@@ -811,11 +803,6 @@ func (r BinaryReader) postingsOffset(name string, values ...string) ([]index.Ran
 }
 
 func (r *BinaryReader) LookupSymbol(o uint32) (string, error) {
-	cacheIndex := o % valueSymbolsCacheSize
-	if cached := r.valueSymbols[cacheIndex]; cached.index == o && cached.symbol != "" {
-		return cached.symbol, nil
-	}
-
 	if s, ok := r.nameSymbols[o]; ok {
 		return s, nil
 	}
@@ -826,16 +813,10 @@ func (r *BinaryReader) LookupSymbol(o uint32) (string, error) {
 		o += headerLen - index.HeaderLen
 	}
 
-	s, err := r.symbols.Lookup(o)
-	if err != nil {
-		return s, err
-	}
-	r.valueSymbols[cacheIndex].index = o
-	r.valueSymbols[cacheIndex].symbol = s
-	return s, nil
+	return r.symbols.Lookup(o)
 }
 
-func (r BinaryReader) LabelValues(name string) ([]string, error) {
+func (r *BinaryReader) LabelValues(name string) ([]string, error) {
 	if r.indexVersion == index.FormatV1 {
 		e, ok := r.postingsV1[name]
 		if !ok {
@@ -891,7 +872,7 @@ func yoloString(b []byte) string {
 	return *((*string)(unsafe.Pointer(&b)))
 }
 
-func (r BinaryReader) LabelNames() ([]string, error) {
+func (r *BinaryReader) LabelNames() ([]string, error) {
 	allPostingsKeyName, _ := index.AllPostingsKey()
 	labelNames := make([]string, 0, len(r.postings))
 	for name := range r.postings {