@yeya24 2021-01-07
The benchmark results are below and I did it twice. Comparing this pr with https://github.com/thanos-io/thanos/pull/3705.

The performance is similar overall, only a little bit improvement.

```
name                                                                 old time/op    new time/op    delta
TelemeterRealData_Series/alerts2w/01EPXBGA0413QZMTF4XQ0M4Q3E            10.8s ± 0%     10.6s ± 0%  -2.12%
TelemeterRealData_Series/alerts15s/01EPXBGA0413QZMTF4XQ0M4Q3E           6.81s ± 0%     6.72s ± 0%  -1.30%
TelemeterRealData_Series/subssyncs2w/01EPXBGA0413QZMTF4XQ0M4Q3E         1.16s ± 0%     1.12s ± 0%  -3.55%
TelemeterRealData_Series/subs2w/01EPXBGA0413QZMTF4XQ0M4Q3E              6.43s ± 0%     5.79s ± 0%  -9.96%
TelemeterRealData_Series/subs2w1account/01EPXBGA0413QZMTF4XQ0M4Q3E      2.00s ± 0%     1.81s ± 0%  -9.20%
TelemeterRealData_Series/subs15m1account/01EPXBGA0413QZMTF4XQ0M4Q3E     2.02s ± 0%     1.87s ± 0%  -7.30%
TelemeterRealData_Series/subs15s/01EPXBGA0413QZMTF4XQ0M4Q3E             3.35s ± 0%     3.21s ± 0%  -4.36%

name                                                                 old alloc/op   new alloc/op   delta
TelemeterRealData_Series/alerts2w/01EPXBGA0413QZMTF4XQ0M4Q3E           4.18GB ± 0%    4.17GB ± 0%  -0.27%
TelemeterRealData_Series/alerts15s/01EPXBGA0413QZMTF4XQ0M4Q3E          1.84GB ± 0%    1.84GB ± 0%  +0.40%
TelemeterRealData_Series/subssyncs2w/01EPXBGA0413QZMTF4XQ0M4Q3E         444MB ± 0%     444MB ± 0%  +0.01%
TelemeterRealData_Series/subs2w/01EPXBGA0413QZMTF4XQ0M4Q3E             2.33GB ± 0%    2.33GB ± 0%  +0.00%
TelemeterRealData_Series/subs2w1account/01EPXBGA0413QZMTF4XQ0M4Q3E      662MB ± 0%     662MB ± 0%  +0.01%
TelemeterRealData_Series/subs15m1account/01EPXBGA0413QZMTF4XQ0M4Q3E     662MB ± 0%     662MB ± 0%  +0.00%
TelemeterRealData_Series/subs15s/01EPXBGA0413QZMTF4XQ0M4Q3E             858MB ± 0%     858MB ± 0%  +0.00%

name                                                                 old allocs/op  new allocs/op  delta
TelemeterRealData_Series/alerts2w/01EPXBGA0413QZMTF4XQ0M4Q3E            34.6M ± 0%     34.7M ± 0%  +0.07%
TelemeterRealData_Series/alerts15s/01EPXBGA0413QZMTF4XQ0M4Q3E           21.5M ± 0%     21.4M ± 0%  -0.12%
TelemeterRealData_Series/subssyncs2w/01EPXBGA0413QZMTF4XQ0M4Q3E         2.41M ± 0%     2.41M ± 0%  +0.01%
TelemeterRealData_Series/subs2w/01EPXBGA0413QZMTF4XQ0M4Q3E              13.8M ± 0%     13.8M ± 0%  +0.00%
TelemeterRealData_Series/subs2w1account/01EPXBGA0413QZMTF4XQ0M4Q3E      2.98M ± 0%     2.98M ± 0%  +0.00%
TelemeterRealData_Series/subs15m1account/01EPXBGA0413QZMTF4XQ0M4Q3E     2.98M ± 0%     2.98M ± 0%  +0.00%
TelemeterRealData_Series/subs15s/01EPXBGA0413QZMTF4XQ0M4Q3E             7.79M ± 0%     7.79M ± 0%  -0.00%
```

```
name                                                                 old time/op    new time/op    delta
TelemeterRealData_Series/alerts2w/01EPXBGA0413QZMTF4XQ0M4Q3E            10.8s ± 0%     10.3s ± 0%   -4.31%
TelemeterRealData_Series/alerts15s/01EPXBGA0413QZMTF4XQ0M4Q3E           6.81s ± 0%     7.35s ± 0%   +7.86%
TelemeterRealData_Series/subssyncs2w/01EPXBGA0413QZMTF4XQ0M4Q3E         1.16s ± 0%     1.06s ± 0%   -8.84%
TelemeterRealData_Series/subs2w/01EPXBGA0413QZMTF4XQ0M4Q3E              6.43s ± 0%     5.70s ± 0%  -11.37%
TelemeterRealData_Series/subs2w1account/01EPXBGA0413QZMTF4XQ0M4Q3E      2.00s ± 0%     1.75s ± 0%  -12.38%
TelemeterRealData_Series/subs15m1account/01EPXBGA0413QZMTF4XQ0M4Q3E     2.02s ± 0%     1.77s ± 0%  -12.08%
TelemeterRealData_Series/subs15s/01EPXBGA0413QZMTF4XQ0M4Q3E             3.35s ± 0%     3.31s ± 0%   -1.46%

name                                                                 old alloc/op   new alloc/op   delta
TelemeterRealData_Series/alerts2w/01EPXBGA0413QZMTF4XQ0M4Q3E           4.18GB ± 0%    4.15GB ± 0%   -0.58%
TelemeterRealData_Series/alerts15s/01EPXBGA0413QZMTF4XQ0M4Q3E          1.84GB ± 0%    1.83GB ± 0%   -0.51%
TelemeterRealData_Series/subssyncs2w/01EPXBGA0413QZMTF4XQ0M4Q3E         444MB ± 0%     444MB ± 0%   +0.01%
TelemeterRealData_Series/subs2w/01EPXBGA0413QZMTF4XQ0M4Q3E             2.33GB ± 0%    2.33GB ± 0%   +0.00%
TelemeterRealData_Series/subs2w1account/01EPXBGA0413QZMTF4XQ0M4Q3E      662MB ± 0%     662MB ± 0%   +0.00%
TelemeterRealData_Series/subs15m1account/01EPXBGA0413QZMTF4XQ0M4Q3E     662MB ± 0%     662MB ± 0%   +0.00%
TelemeterRealData_Series/subs15s/01EPXBGA0413QZMTF4XQ0M4Q3E             858MB ± 0%     780MB ± 0%   -9.11%

name                                                                 old allocs/op  new allocs/op  delta
TelemeterRealData_Series/alerts2w/01EPXBGA0413QZMTF4XQ0M4Q3E            34.6M ± 0%     34.7M ± 0%   +0.05%
TelemeterRealData_Series/alerts15s/01EPXBGA0413QZMTF4XQ0M4Q3E           21.5M ± 0%     21.5M ± 0%   -0.06%
TelemeterRealData_Series/subssyncs2w/01EPXBGA0413QZMTF4XQ0M4Q3E         2.41M ± 0%     2.41M ± 0%   +0.01%
TelemeterRealData_Series/subs2w/01EPXBGA0413QZMTF4XQ0M4Q3E              13.8M ± 0%     13.8M ± 0%   +0.00%
TelemeterRealData_Series/subs2w1account/01EPXBGA0413QZMTF4XQ0M4Q3E      2.98M ± 0%     2.98M ± 0%   +0.00%
TelemeterRealData_Series/subs15m1account/01EPXBGA0413QZMTF4XQ0M4Q3E     2.98M ± 0%     2.98M ± 0%   +0.00%
TelemeterRealData_Series/subs15s/01EPXBGA0413QZMTF4XQ0M4Q3E             7.79M ± 0%     7.79M ± 0%   -0.00%

```

Ideally, this can be improved more by using more fine-grained locks, but I haven't benchmarked it.